Adds chi squared module
This is the class described in #4, and provides the functionality required
by #2, #5 and #6, although the structure is modified. As such this pull request
fixes #2, closes #4, resolves #5 and fixes #6.

The module contains three functions to calculate chi squared for a number given
number of observed events and number of expected events. The forms of chi
squared calculation included are Pearson's, Neyman's and the Poisson Likelihood
chi squared.

The module also adds the `ChiSquared` class that acts as a calculator for a
given `form` of chi squared calculation. It can also be initialised with
penalty term information via the keyword argument `penalty_term`. Or this can
be specified when calling the `get_chi_squared` method. The method takes two 1D
numpy arrays, representing the observed (or "data") spectrum and the expected
(or "MC") spectrum, as arguments and iterates over the array, calculating the
chi squared for each bin using the appropriate function, and adding a penalty
term if required. The method returns the total chi squared comparing the two
spectra.