Modifies chi squared calculations to use `numpy.sum` statements
Performing the arithmetic operations on numpy arrays, inside the `numpy.sum`
method, allows the numpy routines to take care of the iteration, instead of
requiring a python for loop. This proves much faster, especially for large
arrays.

This also adds an error checking function - `check_bin_content` - that checks
all bins in a nupmy array have content >= 0. Each of the chi squared functions
now check the contents of the `observed` and `expected` arrays before
performing the calculation, and raise `ValueError` if the bin contents are not
appropriate.