Adds `zero_nu_limit.py` an example limit setting script.
The script includes examples of how to use the limit setting tools. It also
gives examples of how to implement the key assumptions that accompany spectral
plots, e.g. making a fiducial volume cut or including a region of interest.

Also adds `echidna.__echidna_home__`, which is set to the base directory of the
echidna package in `__init__.py`.