Modifies `limit_config`
Adds functionality to `limit_config` to store the scaling count at the limit
if this is a background/systematic, it saves the count at the chi-squared
minimum as the limit count. If it is a signal it save the scaling count at the
first bin above. Adds function `get_limit_count` to retrieve this value.

Also adds `get_last_bin_above` method, to be used if signal count values are
defined in descending order.