Adds improved ROI handling to Spectra class
Any Region of Interest (ROI) cut can now be made through the `shrink_to_roi`
method. This shrinks the `Spectra` instance but also documents the ROI in the
in the `_rois` attribute of the `Spectra` class. ROIs can be defined and
applied to any axis. The `_rois` dict documents the upper and lower bounds,
plus the efficiency of every ROI cut made through this method.

Also updates `decay` and `limit_setting` modules to use the new ROI handling.