Adds correct `_num_decays` initialisation to parts 2 & 3 of `zero_nu_limit`
Where the spectra are being loaded again from hdf5 for the 2nd and 3rd parts of
the script, they were not being re-initiallised with the correct values for
`_num_decays`, hence the scaling was completely meaningless.

Also fixes a small bug in scaling for `PDFPages` debug plots and some general
tidying up.