Closes #82
Continuing, for now, with the assumption that KamLAND-Zen has the same expected
solar rate as SNO+, updates the `B8_Solar` prior to be the expected rate for the
full 0-15 MeV range from SNO+-doc-507. Note this number has been updated in
SNO+-doc-507v27 for the whitepaper, this commit retains the previous version,
for the #76 verification. The new rate will be included in the next commit.