Closes #92.
I think the problem with the requirements file, was coming from an
old version of SciPy. I've upgraded to the latest version of SciPy.
Also installing both SciPy and numpy was causing problems. Since
numpy is installed as part of the SciPy package, I've removed the
separate install of numpy and we should just use the version that
is installed with the current SciPy version.