Closes #103 and closes #110.
Adds `FitResults` class, which acts as a container for the parmeter grid.
Also contains a copy of the `FitConfig` class, so that it can assign the
best fit values of each `FitParameter`, conce the fit has run.

Also adds the `GridSearch` class - which contains the original grid-
search algorithm. This class inherits from `FitResults` as the
`_data` attribute in `FitResults` is the grid that should be filled
during the grid-search.

Also adds the `_funct` method to `Fit`, which is the begginnings of the
callable that should be passed to any optimiser.