Makes floating with no penalty more intuitive.
Rather than setting sigma as a very large number, this gives the user the
option to completely disable penalty terms for a particular fit parameter, by
explicitly setting sigma as None.

Also improves logging some more and adds a default output path to the `output`
module, which can be overridden to redirect all fitting/limit outputs to a
preferred location. This is very useful if you want to collect all the outputs
of a fit in one place.