Adds logscale_deviation flag to `FitParameter`.
This is an additional scale flag, like `_logscale` that can be used to produce
an array of values - centred around the prior - whose deviations from the prior
are linearly distributed in log-space, so increase exponentially (and
symmetrically) in linear-space.

Also adds `get_bin_boundaries` method to `FitParameter` to achieve same
functionality as `SpectraParameter` method of the same name.