Tidies `spectra` and `limit` modules.
Quite a lot of changes here but all has been thoroughly tested.

The `spectra` module was getting a bit bloated, with all the `Config` and
`Parameter` classes, that shouldn't necessarily have been there.
 * Moves `Config` and all derived classes to `core.config` module
 * Moves `Parameter` and all derived classes to `core.parameter` module
 * Fixes all relative imports in other modules
 * Fixes class and module references in documentation

The `limit` module was also getting to be a bit of a mess. It contained some
modules that have now been superceded by the new fitting/limit setting code.
And the new code that is purely to do with fitting, should really be in a new
`fit` module, not in `limit`.
 * Creates new `fit` module
 * Moves `fit`, `fit_results`, `minimise` and `test_statistic` to `fit` module
 * Fixes relative imports in other files

In `scripts` there were some scripts that now do not work or have been
superceded by newer scripts.
 * Deleted old majoron scripts that do not work
 * `zero_nu_limit` needs updating, marked that it currently doesn't work
 * As well as fixing imports in `getting_started` notebook, updated it a little
   and updated python version accordingly

In `output`, some of the plotting scripts no longer work, but should be
updated:
 * `plot_chi_squared` relies on the (now deleted) `SystAnalyser` class. This
   should be updated to work with the `fit_results` class (both matplotlib and
   ROOT versions), but this is not a priority at the moment, so I've just
   marked that it doesn't work currently.

In `tests`, some unittests are now no longer required as they relate to deleted
modules:
 * Deletes `test_chi_squared` and `test_limit_setting`.
 * `test_chi_squared` already has a replacement in `test_test_statistic`
 * `test_limit_setting` currently has no replacement, we will need to implement
   some form of test of the `Fit` and `Limit` classes.