Fixes stats bug in `Summary`.
Summary class was often not updating the values for `min_per_bin` and using a
previous definition. This often resulted in subtracting the wrong per_bin
values when calculating delta-test-statistic and eneded up with negative values
for the test statistic. This fixes the problem by *always* redefining
`min_per_bin`.