Modifies `klz_majoron_limits` to accept spectra via argparse.
Still, for the majority of cases, spectra will be loaded from the location
specified in the config file. But this gives you the option to parse floating
backgrounds and signals (*Spectra*) directly, for the occasions where you may
wish to modify/manipulate the spectrum first before starting the main script.