Utilises `warnings` module.
In `echidna.utilities` adds `captureWarnings` to logging. This means that
warnings generated through python's `warnings` module are written to the log.
I'm encouraging use of the `warnings` module specifically the `UserWarning`
exception class, for instances where there is something the user could do to
improve the chances of running until completion, but there is also a chance of
running fine to completion with no changes.

Linked to this, I'm adding a `settings` file that can be edited to set echidna-
wide user preferances. The settings file includes a `user_profile` dictionary
that contains a "mode" option - setting this as "cautious" will raise all
`UserWarning`s as exceptions. Otherwise they will not be raised, just logged.

Also, tweaks log colours to make warnings and errors stand out more.