Fixes `logscale_deviation` scale in `FitParameter`.
Improves the algorithm for creating a `logscale_deviation` scale for
`FitParameter`s. The new algorithm still builds an array of values that is
symetric about the prior, with deviations from the prior linear in logspace.
However the values above and below the prior are now treated separately, so it
is possible to build arrays that extend further on one side, preventing the
need to deal with fully symetric arrays extending beyond the bounds for a
parameter type, e.g. rate parameter values extending below zero.

Also updates `check_values` to raise `UserWarning` instead of an exception.
And updates checks.

Marks `get_bin_boundaries` with `PendingDeprecationWarning` as we these are
really just points on a grid, not bins, so shouldn't need boundaries.